requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
gTTS==2.5.3
idna==3.10
numpy==2.1.2
packaging==24.1
pillow==10.4.0
playsound==1.3.0
pytesseract==0.3.13
requests==2.32.3
setuptools==75.1.0
urllib3==2.2.3
wheel==0.44.0
